+++ lib/work_file_lib.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 import os
-from utils import (check_and_create_dir, JsonFormat)
+from lib.utils_lib import (check_and_create_dir, JsonFormat)
 from config.work_file import (DEFAULT_WORK_DIR, WORK_DIR_DICT, WORK_FILE_DICT)
